+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Throwing errors in middleware functions.
 - RestHapi.logger and request.logger are now RestHapi.Log and request.Log
+- Replaced config.generateScopes with config.generateRouteScopes
 
 ## [0.43.0] - 2018-03-23
 ### Changed